+++ snacky-backend/snacky-backend/src/main/java/ca/usherbrooke/fgen/api/service/product_template_Service.java
@@ -143,8 +143,15 @@ public class product_template_Service {
             System.out.println("This is what i am sending to Clovis: ProductTemplate/Delete");
             System.out.println(objectMapper.writeValueAsString(info));
 
-            Mapper.deleteProductTemplate(info);
-            System.out.println("Finished deleting: " + ID);
+            if(Mapper.isTemplateUsed(info))
+            {
+                Mapper.deleteProductTemplate(info);
+                System.out.println("Finished deleting: " + ID);
+            }
+            else
+            {
+                throw new Exception("Template is used, you cannot delete this template");
+            }
         }
         catch (Exception e)
         {
